refactor: 장바구니에 추가 모달 네이밍 변경
