+++ src/components/product/index.js
@@ -1,3 +1,3 @@
-export { default as CartAdd } from 'components/product/CartAdd/CartAdd';
+export { default as CartAddForm } from 'components/product/CartAddForm/CartAddForm';
 export { default as ProductCard } from 'components/product/ProductCard/ProductCard';
 export { default as ProductList } from 'components/product/ProductList/ProductList';
